+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Refined the comma representation of large numbers abbreviated with k, m, b
+- Dungeon Party Finder: Parties with specific classes can now *always* be marked as 'unideal' (additionally to the already existing option to mark a party when 2+ members use the same specific class)
+- Dungeon Performance Overlay: added an alternative text border option
 
 ### Fixed
 - Fixed issue with 'no dung class selected'